+++ changelog.md
@@ -1,1 +1,1 @@
-* Change license.
+* Change accuracy attack the GUI.